pyproject.toml
[tool.poetry]
name = "fullstack16"
version = "0.1.0"
description = ""
authors = ["Alisher <alisherxujanov163@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-bootstrap5 = "^24.3"
django-cleanup = "^9.0.0"
pillow = "^11.1.0"
django-allauth = "^65.3.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
